[clean] hslとyuvをまとめた
