+++ src/lib.rs
@@ -1,4 +1,3 @@
 pub mod fastimg;
 pub mod fast3x3;
-pub mod yuv;
-pub mod hsl;
+pub mod colors;
